Lenard's slpSUSTAIN instruction update tweaked:
(1) Referring to a github.com issue that was closed did not seem good for
self-containment of package, and also the issue log was long and not that
informative. If there is important information for the user here, please find a
way to summarise it concisely within the package.

(2) The clarification of w=NA was a bit grammatically odd. I changed it to
match the phrasing of cluster=NA