Fix nosof94sustain (if broken)
Simulation was originally published under old slpSUSTAIN default of ties =
'first'. I add this explicitly to the function to preserve bacward
compatibility and checked the output matched the previous CRAN version. It does
(it's not viable to have an automated test for simulations in the package
because they take too long to run).