Further expand .Rd
The behaviour of the bias parameter is perhaps not what one
would expect in the case of a multi-outcome network (I think I
would have expected to be able to pass a vector of biases in this
case). No need to implement, but important to document the
behaviour, which is what I've done.